sprites: split sprites.py out of display.py
